a program that prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int count;
-	unsigned fbn1, fbn2, sum;
+	unsigned long fbn1, fbn2, sum;
 	
 	fbn1 = 0;
 	fbn2 = 1;
